fixed my part norm (almost)

+++ src/objects.c
@@ -6,13 +6,13 @@
 /*   By: bel-amri <clorensunity@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/04 21:16:08 by bel-amri          #+#    #+#             */
-/*   Updated: 2023/03/04 21:19:21 by bel-amri         ###   ########.fr       */
+/*   Updated: 2023/03/17 04:33:52 by bel-amri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <header.h>
 
-t_pos	new_pos(int x, int y)
+t_pos	new_pos(float x, float y)
 {
 	t_pos	pos;
